14/02/2021 - Form GUI
13/02/2021{
	NeuralNet.cs -> Modify Name from 'Program.cs' This is  so new program cs file can run the form object
	NeuralNet.cs -> Add static bool 'LastWasApple' to be referenced if the N.N's decision of an image is needed.
         NeuralNet.cs->Main()-> Modify Method to be callable method 'Run Neural Net'
	NeuralNet.cs->RunNeuralNet()-> Modify Comments to match
	NeuralNet.cs->RunNeuralNet() // NeuralNetForm.cs-> Modify Testing section of the method to be run on button call in NeuralNetForm
	NeuralNet.cs->RunNeuralNet()-> Move 'testLayer' to be a static public var so 'TestImage()' can be run using it.



	NeuralNetForm.cs -> Add a Form object. This is to have a presentation GUI of the NeuralNet at work.
		The intention is it will be able to interate uses of the old console-based 'NeuralNet' program
		by making the 'NeuralNet' program a static class that can be called from a form, using a buton and
		textbox object to provide values and call the test section through a button press
	NeuralNetForm.cs -> OnPaint() -> Add overriding onPaint Method to draw last image searched and a cross or tick
		depending on whether the neural net thought it was an apple. This is for visual feedback of the system working.


}
14/02/2021{
	NeuralNet.cs->RunNeuralNet()-> Add right Count var for each correct guess program makes with test data.
	If more than 1 test data is guessed wrong, reruns program. 


}
