+++ ImageReader.cs
@@ -2,8 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Text;
-using static NeuralNet.Program;
-
+using static NeuralNet.NeuralNet;
 
 /// <summary>
 ///
