+++ Layer.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using static NeuralNet.Program;
-
+using static NeuralNet.NeuralNet;
 
 /// <summary>
 ///
